Fix multiple compilation errors and deprecation warnings
- Add ApiResponse.error factory to api_response.dart
- Remove duplicate displayName getter in activity_model.dart
- Replace deprecated WillPopScope with PopScope in app_utils.dart and edit_event_screen.dart
- Rename StringExtension to StringValidationExtension to avoid conflict with GetX
- Fix EventTagsDisplay to EventTags in event_details_screen.dart
- Fix WeatherBadge temperature type from String to int
- Add showShareOptions method to event_details_screen.dart
- Fix num to double type conversion in event_analytics.dart
- Add EventTagsInput widget to event_tags.dart

+++ lib/widgets/event_analytics.dart
@@ -21,8 +21,8 @@ class EventPerformance extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final conversionRate = views > 0 ? (rsvps / views * 100) : 0;
-    final attendanceRate = rsvps > 0 ? (checkins / rsvps * 100) : 0;
+    final conversionRate = views > 0 ? (rsvps / views * 100).toDouble() : 0.0;
+    final attendanceRate = rsvps > 0 ? (checkins / rsvps * 100).toDouble() : 0.0;
 
     return Container(
       padding: const EdgeInsets.all(AppSpacing.md),
